Revert "Teacher/resource"
