+++ app/src/main/java/com/example/project_skripsi/core/repository/FireRepository.kt
@@ -145,6 +145,7 @@ class FireRepository : OnSuccessListener<Void>, OnFailureListener {
     }
 
     fun addTaskForm(taskForm: TaskForm, teacher: Teacher) {
+
         db.collection(COLLECTION_TASK_FORM)
             .document(taskForm.id!!)
             .set(taskForm)
@@ -155,26 +156,6 @@ class FireRepository : OnSuccessListener<Void>, OnFailureListener {
             .set(teacher)
             .addOnSuccessListener(this)
             .addOnFailureListener(this)
-    }
-
-    fun addResource(resource: Resource, teacher: Teacher): Pair<LiveData<Boolean> , LiveData<Exception>> {
-        val isSuccess = MutableLiveData<Boolean>()
-        val exception = MutableLiveData<Exception>()
-        db.collection(COLLECTION_RESOURCE)
-            .document(resource.id!!)
-            .set(resource)
-            .addOnSuccessListener(this)
-            .addOnFailureListener(this)
-        db.collection(COLLECTION_TEACHER)
-            .document(teacher.id!!)
-            .set(teacher)
-            .addOnSuccessListener {
-                isSuccess.value = true
-            }
-            .addOnFailureListener {
-                exception.value = it
-            }
-        return Pair(isSuccess, exception)
     }
 
     override fun onSuccess(p0: Void?) {
