+++ app/src/main/java/com/example/project_skripsi/core/repository/dummy/FirestoreDummy.kt
@@ -99,7 +99,7 @@ class FirestoreDummy : OnSuccessListener<Any>, OnFailureListener {
                         "Biologi",
                         12,
                         listOf("eMsulnik6kEpW0ESKI9V"),
-                        mutableListOf("dxXTXZcrj0yVh8PpzYk2", "feiaZB0ds1rbaWT1g8hJ", "RS1201", "RS1101", "RF1201"),
+                        listOf("dxXTXZcrj0yVh8PpzYk2", "feiaZB0ds1rbaWT1g8hJ", "RS1201", "RS1101", "RF1201"),
                         mutableListOf("HaWuFgmvLAuZYeG5JuVw"),
                         mutableListOf("ripyBsBZObBfarZpd085")
                     )
