+++ app/src/main/java/com/example/project_skripsi/module/teacher/form/alter_task/TcAlterTaskFragment.kt
@@ -2,7 +2,6 @@ package com.example.project_skripsi.module.teacher.form.alter_task
 
 import android.annotation.SuppressLint
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -136,8 +135,10 @@ class TcAlterTaskFragment : Fragment() {
         }
 
         retrieveArgs()
+
         return binding.root
     }
+
 
 
     override fun onDestroyView() {
