+++ app/src/main/java/com/example/project_skripsi/module/teacher/main/task/TcTaskFragment.kt
@@ -29,13 +29,14 @@ class TcTaskFragment : Fragment(), ItemClickListener {
 
         viewModel = ViewModelProvider(this)[TcTaskViewModel::class.java]
         _binding = FragmentTcTaskBinding.inflate(inflater, container, false)
+
         viewModel.subjectGroupList.observe(viewLifecycleOwner, {
             binding.cgSubjectGroup.removeAllViews()
             var hasItem = false
             it.map { subjectGroup ->
                 val chip = inflater.inflate(R.layout.standard_chip_choice, binding.cgSubjectGroup, false) as Chip
                 chip.id = View.generateViewId()
-                chip.text = ("${subjectGroup.gradeLevel} - ${subjectGroup.subjectName}")
+                chip.text = ("${subjectGroup.subjectName} - ${subjectGroup.gradeLevel}")
                 chip.setOnCheckedChangeListener { _, isChecked ->
                     if (isChecked) viewModel.selectSubjectGroup(subjectGroup)
                 }
