+++ app/src/main/java/com/example/project_skripsi/module/teacher/main/task/TcTaskViewModel.kt
@@ -11,6 +11,7 @@ import com.example.project_skripsi.utils.generic.GenericObserver.Companion.obser
 
 class TcTaskViewModel : ViewModel() {
 
+
     private val _subjectGroupList = MutableLiveData<List<SubjectGroup>>()
     val subjectGroupList : LiveData<List<SubjectGroup>> = _subjectGroupList
