--- app/src/main/res/layout/item_tc_resource.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:padding="8dp"
-    android:layout_margin="4dp"
-    app:cardCornerRadius="6dp">
-
-    <TextView
-        android:id="@+id/tvResourceTitle"
-        android:layout_width="wrap_content"
-        android:layout_height="match_parent"
-        android:padding="8dp"
-        android:text="Materi 1"/>
-
-</androidx.cardview.widget.CardView>
