--- app/src/main/res/layout/tc_item_chip.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<com.google.android.material.chip.Chip style="@style/Widget.MaterialComponents.Chip.Choice"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:text="@string/chip_text"
-    app:chipIconEnabled="true"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto" />
